webapp/requirements.txt
flask>=0.10.1
flask_debugtoolbar
psycopg2
sqlalchemy
python-dotenv
flask-sqlalchemy
flask-bootstrap
flask-nav
flask-debug
